requirements.txt
openai
anthropic
google-generativeai
requests
SPARQLWrapper
pyyaml
python-dotenv
pandas
tqdm
jsonlines
google-genai
dashscope
